Add code and Unit test for the AI Driver code
This is simplified AI. Its written using generators which makes it
easier to mix different responsibilities with callbacks to maintain
state. Like the conversation context is maintained in AIDriver. History
and Human In loop convo is maintained by the driver_wrapper. The client
code is much simpler: It only has to worry about coding in some of the
call backs.